refactor: move new-workspace-enter-guard to src/renderer/src/lib
