+++ src/renderer/src/lib/new-workspace-enter-guard.test.ts
