+++ src/renderer/src/lib/new-workspace-enter-guard.ts
